+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 0.2 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Polyfill removeDuplicateObjects so that we can use version 2.0.13 of
+  Patternslib with Plone/Mockup which still uses Patternslib 2.0.11.
 
 
 0.1 (2016-04-27)